Refactor: ts마이그레이션 (auth 미들웨어)
